Forge/Neo fixes
