+++ forge/src/main/java/gjum/minecraft/civ/snitchmod/forge/mixins/LevelRenderMixin.java
@@ -1,9 +1,15 @@
 package gjum.minecraft.civ.snitchmod.forge.mixins;
 
+import com.mojang.blaze3d.buffers.GpuBufferSlice;
+import com.mojang.blaze3d.resource.GraphicsResourceAllocator;
 import com.mojang.blaze3d.vertex.PoseStack;
 import gjum.minecraft.civ.snitchmod.forge.ForgeSnitchMod;
+import net.minecraft.client.Camera;
+import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.renderer.LevelRenderer;
+import net.minecraft.client.renderer.rendertype.RenderType;
 import org.joml.Matrix4f;
+import org.joml.Vector4f;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -11,10 +17,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(LevelRenderer.class)
 public abstract class LevelRenderMixin { // REMINDER: Forge sucks
-    /*@Inject(method = "renderSectionLayer", at = @At("RETURN"))
-    private void onRenderLevelLast(RenderType arg, double d, double e, double f, Matrix4f matrix4f, Matrix4f matrix4f2, CallbackInfo ci) {
-        if (arg == RenderType.translucent()) {
-            ForgeSnitchMod.getMod().handleRenderBlockOverlay(new PoseStack().last().pose());
-        }
-    }*/
+
+    @Inject(method = "method_62214", at = @At(value = "INVOKE:LAST", target = "Lnet/minecraft/client/renderer/MultiBufferSource$BufferSource;endBatch()V"))
+    private void onRenderLevelLast(CallbackInfo ci) {
+        ForgeSnitchMod.getMod().handleRenderBlockOverlay(new PoseStack());
+    }
 }
